requirements.txt
beautifulsoup4>=4.12.0
PyYAML>=6.0
requests>=2.28.0
